+++ CHANGELOG.rst
@@ -8,6 +8,8 @@ Changelog
   attempting to process the message (workaround for servers not able
   to process certain requests, maybe a nicer generic soultion can be
   found or it can be made configurable when starting the server)
+- Fix exception handling in the example exception logger middleware
+- Major update of the aio_pika example server code
 
 1.4.2 (2022-04-14)
 ------------------